+++ lib/src/transaction/constants.dart
@@ -5,3 +5,6 @@
 /// 40 bytes is the size of the IPv6 header
 /// 8 bytes is the size of the fragment header
 const packetDataSize = 1280 - 40 - 8;
+
+/// Version message number prefix mask.
+const versionPrefixMask = 0x7f;