+++ lib/src/utils/shortvec.dart
@@ -1,16 +1,25 @@
+/// Imports
+/// ------------------------------------------------------------------------------------------------
+
+import 'package:solana_common/utils/buffer.dart';
+
+
 /// Shortvec Encoding
 /// ------------------------------------------------------------------------------------------------
 
-/// Returns the short-vec [encoded] length.
-int decodeLength(final Iterable<int> encoded) {
+/// Returns the short-vec encoded length.
+int decodeLength(final Buffer buffer) {
   int len = 0;
   int size = 0;
-  for (final int byte in encoded) {
+  for (final int byte in buffer) {
     len |= (byte & 0x7f) << (size * 7);
     size += 1;
     if ((byte & 0x80) == 0) {
       break;
     }
+  }
+  if (size > 0) {
+    buffer.set(buffer.getRange(size).toList(growable: false));
   }
   return len;
 }